+++ dd-java-agent/instrumentation/vertx-pg-client-4.2.0/src/main/java/datadog/trace/instrumentation/vertx_pg_client_4_2_0/PgConnectionFactoryConstructorAdvice.java
@@ -3,7 +3,6 @@ package datadog.trace.instrumentation.vertx_pg_client_4_2_0;
 import datadog.trace.bootstrap.InstrumentationContext;
 import datadog.trace.bootstrap.instrumentation.jdbc.DBInfo;
 import io.vertx.pgclient.PgConnectOptions;
-import io.vertx.pgclient.PgConnection;
 import io.vertx.pgclient.impl.PgConnectionFactory;
 import net.bytebuddy.asm.Advice;
 
@@ -22,10 +21,5 @@ public class PgConnectionFactoryConstructorAdvice {
             .type("postgresql")
             .build();
     InstrumentationContext.get(PgConnectionFactory.class, DBInfo.class).put(zis, info);
-  }
-
-  // Limit ourselves to 4.x by checking for the ping() method that was added in 4.x
-  private static void muzzleCheck(PgConnection connection) {
-    connection.query("SELECT 1");
   }
 }
