+++ dd-java-agent/instrumentation/vertx-pg-client-4.2.0/src/main/java/datadog/trace/instrumentation/vertx_pg_client_4_2_0/PreparedStatementQueryAdvice.java
@@ -2,7 +2,6 @@ package datadog.trace.instrumentation.vertx_pg_client_4_2_0;
 
 import datadog.trace.api.Pair;
 import datadog.trace.bootstrap.InstrumentationContext;
-import io.vertx.pgclient.PgConnection;
 import io.vertx.sqlclient.PreparedStatement;
 import io.vertx.sqlclient.Query;
 import net.bytebuddy.asm.Advice;
@@ -13,10 +12,5 @@ public class PreparedStatementQueryAdvice {
       @Advice.This final PreparedStatement zis, @Advice.Return final Query query) {
     InstrumentationContext.get(Query.class, Pair.class)
         .put(query, InstrumentationContext.get(PreparedStatement.class, Pair.class).get(zis));
-  }
-
-  // Limit ourselves to 4.x by checking for the ping() method that was added in 4.x
-  private static void muzzleCheck(PgConnection connection) {
-    connection.query("SELECT 1");
   }
 }
