+++ dd-java-agent/instrumentation/vertx-pg-client-4.2.0/src/main/java/datadog/trace/instrumentation/vertx_pg_client_4_2_0/QueryAdvice.java
@@ -1,10 +1,5 @@
 package datadog.trace.instrumentation.vertx_pg_client_4_2_0;
 
-import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.activateSpan;
-import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.activeSpan;
-import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.captureSpan;
-import static datadog.trace.instrumentation.vertx_pg_client_4_2_0.VertxSqlClientDecorator.DECORATE;
-
 import datadog.trace.api.Pair;
 import datadog.trace.bootstrap.ContextStore;
 import datadog.trace.bootstrap.InstrumentationContext;
@@ -12,11 +7,13 @@ import datadog.trace.bootstrap.instrumentation.api.AgentScope;
 import datadog.trace.bootstrap.instrumentation.api.AgentSpan;
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Handler;
-import io.vertx.pgclient.PgConnection;
 import io.vertx.sqlclient.Query;
 import io.vertx.sqlclient.SqlResult;
 import net.bytebuddy.asm.Advice;
 import net.bytebuddy.implementation.bytecode.assign.Assigner;
+
+import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.*;
+import static datadog.trace.instrumentation.vertx_pg_client_4_2_0.VertxSqlClientDecorator.DECORATE;
 
 public class QueryAdvice {
   public static class Copy {
@@ -25,11 +22,6 @@ public class QueryAdvice {
         @Advice.This final Query<?> zis, @Advice.Return final Query<?> ret) {
       ContextStore<Query, Pair> contextStore = InstrumentationContext.get(Query.class, Pair.class);
       contextStore.put(ret, contextStore.get(zis));
-    }
-
-    // Limit ourselves to 4.x by checking for the ping() method that was added in 4.x
-    private static void muzzleCheck(PgConnection connection) {
-      connection.query("SELECT 1");
     }
   }
 
@@ -72,11 +64,6 @@ public class QueryAdvice {
       if (null != clientScope) {
         clientScope.close();
       }
-    }
-
-    // Limit ourselves to 4.x by checking for the ping() method that was added in 4.x
-    private static void muzzleCheck(PgConnection connection) {
-      connection.query("SELECT 1");
     }
   }
 }
